backend/requirements.txt
Django>=5.2.0
djangorestframework>=3.16.0
django-cors-headers>=4.7.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
celery>=5.5.0
redis>=6.2.0
djangorestframework-simplejwt>=5.5.0
Pillow>=11.0.0
pytesseract>=0.3.13
openai>=1.52.0

# Testing dependencies
pytest>=8.3.0
pytest-django>=4.9.0
factory-boy>=3.3.1
responses>=0.25.0
coverage>=7.6.0
locust>=2.32.0